agent/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
annoy==1.17.3
anyio==4.8.0
attrs==25.1.0
av==14.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
distro==1.9.0
flatbuffers==25.2.10
frozenlist==1.5.0
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
jiter==0.9.0
livekit==0.21.3
livekit-agents==0.12.17
livekit-api==0.8.2
livekit-plugins-cartesia==0.4.10
livekit-plugins-deepgram==0.7.0
livekit-plugins-elevenlabs==0.8.0
livekit-plugins-openai==1.0.1
livekit-plugins-rag==0.2.4
livekit-plugins-silero==0.7.5
livekit-protocol==0.9.2
mpmath==1.3.0
multidict==6.1.0
numpy==2.2.3
onnxruntime==1.21.0
openai==1.65.5
packaging==24.2
pillow==11.1.0
pip==25.0
portalocker==2.10.1
propcache==0.3.0
protobuf==5.29.3
psutil==5.9.8
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
PyPDF2==3.0.1
python-dotenv==1.0.1
qdrant-client==1.13.3
regex==2024.11.6
requests==2.32.3
setuptools==75.8.0
sniffio==1.3.1
sympy==1.13.3
tiktoken==0.9.0
tqdm==4.67.1
types-protobuf==4.25.0.20240417
typing_extensions==4.12.2
urllib3==2.3.0
watchfiles==0.24.0
wheel==0.45.1
yarl==1.18.3